Tests/LibWeb/Text/input/Editing/input-event.html

<script src="../include.js"></script>
<style>
    #input {
        width: 100px;
        height: 100px;
        border: 1px solid black;
    }
</style>
<div id="input" contenteditable="true"></div>
<script>
    test(() => {
        const input = document.getElementById("input");
        input.addEventListener("input", (e) => {
            println(`input data=(${e.data}) intputType=(${e.inputType})`);
        });
        internals.sendText(input, "hello");
        internals.commitText();
    });
</script>
