Tests/LibWeb/Text/expected/Editing/input-event.txt

input data=(h) intputType=(insertText)
input data=(e) intputType=(insertText)
input data=(l) intputType=(insertText)
input data=(l) intputType=(insertText)
input data=(o) intputType=(insertText)
input data=(null) intputType=(insertParagraph)
